Rename files to move them in the IDE
